albu-solution/requirements.txt
cython
numpy==1.13.1
opencv-python
pyproj
matplotlib
networkx==1.11
fiona
scikit-image==0.13.1
scikit-learn
scipy==1.0.0
geopandas
tqdm
torchvision
utm
rtree
tensorboardX
numba==0.36.2
pandas==0.21.0
